+++ common/src/main/java/com/faboslav/structurify/common/checks/JigsawStructureBiomeCheck.java
@@ -1,13 +1,13 @@
 package com.faboslav.structurify.common.checks;
 
 import com.faboslav.structurify.common.config.data.StructureData;
-import net.minecraft.core.BlockPos;
 import net.minecraft.core.Holder;
+import net.minecraft.core.HolderSet;
 import net.minecraft.core.QuartPos;
 import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.biome.CheckerboardColumnBiomeSource;
-import net.minecraft.world.level.levelgen.WorldGenerationContext;
+import net.minecraft.world.level.levelgen.Heightmap;
 import net.minecraft.world.level.levelgen.heightproviders.HeightProvider;
 import net.minecraft.world.level.levelgen.structure.Structure;
 
@@ -16,7 +16,8 @@ public final class JigsawStructureBiomeCheck
 	public static boolean checkBiomes(
 		StructureData structureData,
 		HeightProvider startHeight,
-		Structure.GenerationContext generationContext
+		Structure.GenerationContext generationContext,
+		HolderSet<Biome> blacklistedBiomes
 	) {
 		var biomeCheckDistance = (int) Math.ceil(structureData.getBiomeCheckDistance() / 16.0);
 
@@ -24,18 +25,44 @@ public final class JigsawStructureBiomeCheck
 			return true;
 		}
 
-		ChunkPos chunkPos = generationContext.chunkPos();
-		int y = startHeight.sample(generationContext.random(), new WorldGenerationContext(generationContext.chunkGenerator(), generationContext.heightAccessor()));
-		var blockPos = new BlockPos(generationContext.chunkPos().getMinBlockX(), y, generationContext.chunkPos().getMinBlockZ());
+		if(structureData.getBiomeCheckMode() == StructureData.BiomeCheckMode.BLACKLIST && structureData.getBiomeCheckBlacklistedBiomes().isEmpty()) {
+			return true;
+		}
 
-		int sectionY = QuartPos.fromBlock(blockPos.getY());
+		ChunkPos chunkPos = generationContext.chunkPos();
 
 		for (int curChunkX = chunkPos.x - biomeCheckDistance; curChunkX <= chunkPos.x + biomeCheckDistance; curChunkX++) {
 			for (int curChunkZ = chunkPos.z - biomeCheckDistance; curChunkZ <= chunkPos.z + biomeCheckDistance; curChunkZ++) {
-				Holder<Biome> biome = generationContext.biomeSource().getNoiseBiome(QuartPos.fromSection(curChunkX), sectionY, QuartPos.fromSection(curChunkZ), generationContext.randomState().sampler());
-				if (!generationContext.validBiome().test(biome)) {
+
+				int blockX = curChunkX << 4;
+				int blockZ = curChunkZ << 4;
+
+				int surfaceY = generationContext.chunkGenerator().getFirstFreeHeight(
+					blockX,
+					blockZ,
+					Heightmap.Types.WORLD_SURFACE_WG,
+					generationContext.heightAccessor(),
+					generationContext.randomState()
+				);
+
+				Holder<Biome> biome = generationContext.biomeSource().getNoiseBiome(
+					QuartPos.fromBlock(blockX),
+					QuartPos.fromBlock(surfaceY),
+					QuartPos.fromBlock(blockZ),
+					generationContext.randomState().sampler()
+				);
+
+				if (structureData.getBiomeCheckMode() == StructureData.BiomeCheckMode.STRICT && !generationContext.validBiome().test(biome)) {
 					return false;
 				}
+
+				if (structureData.getBiomeCheckMode() == StructureData.BiomeCheckMode.BLACKLIST && blacklistedBiomes.contains(biome)) {
+					return false;
+				}
+
+				// TODO this check will not work for nether because of the Y level checks
+				//BlockPos blockPos = new BlockPos(blockX, surfaceY, blockZ);
+				//Structurify.getLogger().info("Structure biome " + biome + " at surface pos: " + blockPos);
 			}
 		}