+++ common/src/main/java/com/faboslav/structurify/common/config/client/gui/StructurifyConfigScreen.java
@@ -50,9 +50,10 @@ public class StructurifyConfigScreen extends Screen
 	public void render(GuiGraphics context, int mouseX, int mouseY, float delta) {
 		//? <1.20.2 {
 		/*super.renderBackground(context);
-		 *///?} else {
+		 *///?} else if <=1.21.5 {
 		super.renderBackground(context, mouseX, mouseY, delta);
 		//?}
+
 		super.render(context, mouseX, mouseY, delta);
 
 		assert this.minecraft != null;