+++ common/src/main/java/com/faboslav/structurify/common/config/client/gui/widget/ImageButtonWidget.java
@@ -97,12 +97,20 @@ public class ImageButtonWidget extends AbstractWidget
 						float neededWidth = frameWidth * ((float) this.height / frameHeight);
 
 						// Scale the image to fit within the width and height of the button.
+						//? >= 1.21.6 {
+						/*context.pose().pushMatrix();
+						*///?} else {
 						context.pose().pushPose();
+						//?}
 						// gl bilinear scaling.
 						glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 						glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 						contentImage.render(context, getX(), getY(), (int) Math.max(neededWidth, this.width), delta);
+						//? >= 1.21.6 {
+						/*context.pose().popMatrix();
+						*///?} else {
 						context.pose().popPose();
+						 //?}
 
 						// reset gl scaling
 
@@ -139,12 +147,21 @@ public class ImageButtonWidget extends AbstractWidget
 
 		context.fill(unscaledTextX - 5, unscaledTextY - 5, unscaledTextX + this.width - 5, unscaledTextY + client.font.lineHeight + 5, 0xAF000000);
 
+		//? >= 1.21.6 {
+		/*context.pose().pushMatrix();
+		context.pose().scale(fontScaling, fontScaling);
+		*///?} else {
 		context.pose().pushPose();
 		context.pose().scale(fontScaling, fontScaling, 1.0f);
+		 //?}
 
 		renderScrollingString(context, client.font, getMessage(), textX, textY, endX, endY, 0xFFFFFF);
 
+		//? >= 1.21.6 {
+		/*context.pose().popMatrix();
+		*///?} else {
 		context.pose().popPose();
+		 //?}
 
 		// Draw border.
 		context.renderOutline(getX(), getY(), width, height, 0x0FFFFFFF);